feat(CB2-21324): remove old create screen
