+++ package.json
@@ -88,7 +88,7 @@
     "@dvsa/biome-config": "0.2.0",
     "@ngrx/schematics": "^18.0.2",
     "@sentry/cli": "^2.32.2",
-    "@types/jest": "^29.5.12",
+    "@types/jest": "^29.5.14",
     "@types/json-server": "^0.14.7",
     "@types/lodash.clonedeep": "^4.5.9",
     "@types/lodash.merge": "^4.6.9",
