+++ src/app/core/components/header/header.component.html
@@ -1,11 +1,8 @@
 <a [href]="skipLinkHref" class="govuk-skip-link" data-module="govuk-skip-link">Skip to main content</a>
 
 @if (
-  featureToggleService.isFeatureEnabled('betas') && (
-    featureToggleService.isFeatureEnabled('techrecordredesign') ||
-    featureToggleService.isFeatureEnabled('techrecordredesigncreate') ||
-    featureToggleService.isFeatureEnabled('techrecordredesigncreatedetails')
-  )
+  featureToggleService.isFeatureEnabled('betas') && 
+  featureToggleService.isFeatureEnabled('techrecordredesigncreatedetails')
 ) {
   <header class="govuk-header header" role="banner" data-module="govuk-header">
     <div class="govuk-header__container header-container govuk-!-padding-top-4 govuk-!-padding-bottom-2">
