+++ src/app/features/tech-record/create/create-tech-record/__tests__/create-tech-record.component.spec.ts
@@ -1,28 +1,49 @@
 import { provideHttpClient } from '@angular/common/http';
 import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormControl, FormGroup, FormGroupDirective, ReactiveFormsModule } from '@angular/forms';
 import { ActivatedRoute, Router, provideRouter } from '@angular/router';
 import { GlobalErrorService } from '@core/components/global-error/global-error.service';
 
 import { SEARCH_TYPES } from '@models/search-types-enum';
-import { provideMockStore } from '@ngrx/store/testing';
-import { DynamicFormService } from '@services/dynamic-forms/dynamic-form.service';
+import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 
+import { TechRecordType } from '@dvsa/cvs-type-definitions/types/v3/tech-record/tech-record-vehicle-type';
+import { TechRecordType as V3TechRecordType } from '@dvsa/cvs-type-definitions/types/v3/tech-record/tech-record-verb-vehicle-type';
 import { initialAppState } from '@store/index';
 import { of } from 'rxjs';
 import { CreateTechRecordComponent } from '../create-tech-record.component';
 
-describe('CreateNewVehicleRecordComponent', () => {
+describe('CreateTechRecordComponent', () => {
 	let component: CreateTechRecordComponent;
 	let fixture: ComponentFixture<CreateTechRecordComponent>;
 	let errorService: GlobalErrorService;
 	let route: ActivatedRoute;
 	let router: Router;
+	let store: MockStore;
 	let techRecordService: TechnicalRecordService;
+	let formGroupDirective: FormGroupDirective;
+
+	const techRecord: V3TechRecordType<'hgv', 'put'> = {
+		techRecord_vehicleType: 'hgv',
+		partialVin: '',
+		techRecord_bodyType_description: '',
+		techRecord_noOfAxles: 2,
+		techRecord_reasonForCreation: 'test',
+		techRecord_statusCode: 'provisional',
+		techRecord_vehicleClass_description: 'heavy goods vehicle',
+		primaryVrm: '',
+		vin: '',
+		techRecord_adrDetails_dangerousGoods: false,
+	};
 
 	beforeEach(async () => {
+		formGroupDirective = new FormGroupDirective([], []);
+		formGroupDirective.form = new FormGroup<
+			Partial<Record<keyof TechRecordType<'hgv' | 'car' | 'psv' | 'lgv' | 'trl'>, FormControl>>
+		>({});
+
 		await TestBed.configureTestingModule({
 			imports: [CreateTechRecordComponent, ReactiveFormsModule],
 			providers: [
@@ -42,48 +63,13 @@ describe('CreateNewVehicleRecordComponent', () => {
 		route = TestBed.inject(ActivatedRoute);
 		router = TestBed.inject(Router);
 		techRecordService = TestBed.inject(TechnicalRecordService);
+		store = TestBed.inject(MockStore);
 		component = fixture.componentInstance;
 		fixture.detectChanges();
 	});
 
 	it('should create', () => {
 		expect(component).toBeTruthy();
-	});
-
-	describe('get vehicleTypeOptions', () => {
-		it('should return the expected options', () => {
-			expect(component.vehicleTypeOptions).toBeTruthy();
-		});
-	});
-
-	describe('get vehicleStatusOptions', () => {
-		it('should return the expected options', () => {
-			expect(component.vehicleStatusOptions).toBeTruthy();
-		});
-	});
-
-	describe('get isFormValid', () => {
-		it('should call validate with the vehicleForm and an empty array', () => {
-			const validateSpy = jest.spyOn(DynamicFormService, 'validate').mockImplementation();
-			const isValid = component.isFormValid;
-			expect(isValid).toBeDefined();
-			expect(validateSpy).toHaveBeenCalledTimes(1);
-			expect(validateSpy).toHaveBeenCalledWith(component.form, []);
-		});
-
-		it('should call setErrors with an empty array', () => {
-			jest.spyOn(DynamicFormService, 'validate').mockImplementation(() => {});
-			const setErrorsSpy = jest.spyOn(errorService, 'setErrors').mockImplementation();
-			const isValid = component.isFormValid;
-			expect(isValid).toBeDefined();
-			expect(setErrorsSpy).toHaveBeenCalledTimes(1);
-			expect(setErrorsSpy).toHaveBeenCalledWith([]);
-		});
-
-		it('should return vehicleForm.valid', () => {
-			const formValid = component.isFormValid;
-			expect(formValid).toBeFalsy();
-		});
 	});
 
 	describe('navigateBack', () => {
@@ -107,31 +93,13 @@ describe('CreateNewVehicleRecordComponent', () => {
 	});
 
 	describe('handleSubmit', () => {
-		it('should do nothing if the form is not valid', async () => {
-			const formUniqueSpy = jest.spyOn(component, 'isFormValueUnique').mockImplementation();
-			await component.handleSubmit();
-			expect(formUniqueSpy).toHaveBeenCalledTimes(0);
-		});
-
-		it('should do nothing if the form value not unique', () => {
-			const isFormValid = jest.spyOn(component, 'isFormValid', 'get').mockReturnValue(true);
-			const updateEditingSpy = jest.spyOn(techRecordService, 'updateEditingTechRecord');
-			const navigateSpy = jest.spyOn(router, 'navigate');
-			const generateTechREcordSpy = jest.spyOn(
-				techRecordService,
-				'generateEditingVehicleTechnicalRecordFromVehicleType'
-			);
-			void component.handleSubmit();
-
-			expect(isFormValid).toHaveReturned();
-			expect(updateEditingSpy).toHaveBeenCalledTimes(0);
-			expect(generateTechREcordSpy).toHaveBeenCalledTimes(0);
-			expect(navigateSpy).toHaveBeenCalledTimes(0);
+		beforeEach(() => {
+			component.vinUnique = true;
 		});
 
 		it('should navigate to hydrate when successful', async () => {
-			jest.spyOn(component, 'isFormValid', 'get').mockReturnValue(true);
 			jest.spyOn(component, 'isFormValueUnique').mockImplementation(() => Promise.resolve(true));
+			jest.spyOn(errorService, 'extractGlobalErrors').mockImplementation(() => []);
 			const routerSpy = jest.spyOn(router, 'navigate').mockImplementation(() => Promise.resolve(true));
 			jest.spyOn(techRecordService, 'updateEditingTechRecord');
 
@@ -140,10 +108,41 @@ describe('CreateNewVehicleRecordComponent', () => {
 			fixture.detectChanges();
 			expect(routerSpy).toHaveBeenCalledWith(['../create/new-record-details'], { relativeTo: route });
 		});
+
+		it('should set the global errors if form is not valid', async () => {
+			jest.spyOn(errorService, 'extractGlobalErrors').mockImplementation(() => [{ error: 'Form is invalid' }]);
+			jest.spyOn(errorService, 'setErrors');
+
+			await component.handleSubmit();
+
+			expect(errorService.setErrors).toHaveBeenCalled();
+		});
+
+		it('should not navigate anywhere if the VRM or Trailer ID is not unique', async () => {
+			jest.spyOn(component, 'isFormValueUnique').mockImplementation(() => Promise.resolve(false));
+			jest.spyOn(errorService, 'extractGlobalErrors').mockImplementation(() => []);
+			const routerSpy = jest.spyOn(router, 'navigate');
+
+			await component.handleSubmit();
+
+			expect(routerSpy).not.toHaveBeenCalled();
+		});
+
+		it('should navigate to the duplicate VIN page if the VIN is not unique', async () => {
+			jest.spyOn(component, 'isFormValueUnique').mockImplementation(() => Promise.resolve(false));
+			jest.spyOn(errorService, 'extractGlobalErrors').mockImplementation(() => []);
+			const routerSpy = jest.spyOn(router, 'navigate');
+			component.vrmUnique = true;
+			component.vinUnique = false;
+
+			await component.handleSubmit();
+
+			expect(routerSpy).toHaveBeenCalledWith(['../create/duplicate-vin'], { relativeTo: component.route });
+		});
 	});
 
 	describe('isVinUnique', () => {
-		it('should call isUnique with an emptry string and the type of vin', async () => {
+		it('should call isUnique with an empty string and the type of vin', async () => {
 			const isUniqueSpy = jest.spyOn(techRecordService, 'isUnique').mockImplementation(() => of(true));
 
 			await component.isVinUnique();
@@ -169,38 +168,13 @@ describe('CreateNewVehicleRecordComponent', () => {
 			expect(result).toBeTruthy();
 		});
 
-		it('should call addError when the VRM is not unique', async () => {
+		it('should call addError when the VRM must be unique', async () => {
 			jest.spyOn(techRecordService, 'isUnique').mockImplementation(() => of(false));
 			const addErrorSpy = jest.spyOn(errorService, 'addError').mockImplementation();
 
 			const result = await component.isVrmUnique();
 
-			expect(addErrorSpy).toHaveBeenCalledWith({ error: 'Vrm not unique', anchorLink: 'input-vrm-or-trailer-id' });
-			expect(result).toBeFalsy();
-		});
-	});
-
-	describe('isTrailerIdUnique', () => {
-		it('should return true when the trailer ID is unique', async () => {
-			jest.spyOn(techRecordService, 'isUnique').mockImplementation(() => of(true));
-			component.techRecord = { techRecord_vehicleType: 'trl' };
-
-			const result = await component.isTrailerIdUnique();
-
-			expect(result).toBeTruthy();
-		});
-
-		it('should call addError when the trailer ID is not unique', async () => {
-			jest.spyOn(techRecordService, 'isUnique').mockImplementation(() => of(false));
-			const addErrorSpy = jest.spyOn(errorService, 'addError').mockImplementation();
-			component.techRecord = { techRecord_vehicleType: 'trl' };
-
-			const result = await component.isTrailerIdUnique();
-
-			expect(addErrorSpy).toHaveBeenCalledWith({
-				error: 'Trailer ID must be unique',
-				anchorLink: 'input-vrm-or-trailer-id',
-			});
+			expect(addErrorSpy).toHaveBeenCalledWith({ error: 'VRM must be unique', anchorLink: 'input-vrm-or-trailer-id' });
 			expect(result).toBeFalsy();
 		});
 	});
