--- src/app/features/tech-record/create/create-wrapper/create-v1/create-tech-record.component.html
@@ -1,66 +1,0 @@
-<form [formGroup]="form">
-  <h1 class="govuk-heading-l title">New vehicle information</h1>
-
-  <app-text-input
-    formControlName="vin"
-    name="input-vin"
-    label="Vehicle Identification number (VIN)"
-    appToUppercase
-    appNoSpace
-    appTrimWhitespace
-    hint="VIN should usually be 17 characters long."
-    [width]="30"
-    (ngModelChange)="ngOnChanges()"
-  ></app-text-input>
-
-  @if (!generateID.value) {
-    <app-text-input
-      formControlName="vrmTrm"
-      name="input-vrm-or-trailer-id"
-      label="VRM / Trailer ID"
-      appToUppercase
-      appNoSpace
-      appTrimWhitespace
-      [width]="30"
-      (ngModelChange)="ngOnChanges()"
-    ></app-text-input>
-  }
-
-  @if (!generateID.value) {
-    <span class="govuk-list">Or</span>
-  }
-
-  <app-checkbox-group
-    #generateID
-    formControlName="generateID"
-    name="generate-c-t-z-num"
-    [options]="checkboxOptions"
-    (ngModelChange)="toggleVrmInput(generateID)"
-  ></app-checkbox-group>
-
-  <app-select
-    formControlName="vehicleStatus"
-    name="change-vehicle-status-select"
-    label="Select vehicle status"
-    [options]="vehicleStatusOptions"
-    (ngModelChange)="ngOnChanges()"
-  ></app-select>
-
-  <app-select
-    formControlName="vehicleType"
-    name="change-vehicle-type-select"
-    label="Select vehicle type"
-    [options]="vehicleTypeOptions"
-    (ngModelChange)="ngOnChanges()"
-  ></app-select>
-
-  @if (isVinUniqueCheckComplete && !vinUnique) {
-    <p class="govuk-error-message">
-      The Vin you are about to submit is a duplicate, if you are happy to continue please press continue
-    </p>
-  }
-
-  <app-button id="create-record-continue" (clicked)="handleSubmit()">Continue</app-button>
-
-  <app-button design="link" data-module="govuk-button" (clicked)="navigateBack()">Back</app-button>
-</form>
